test/my-spring-boot-app/src/main/java/com/example/myapp/controller/JwtControllerTest.java
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.util.Date;

public class JwtControllerTest {

    @Test
    public void testValidToken() {
        JwtController jwtController = new JwtController();
        String validToken = generateValidToken();

        boolean isValid = jwtController.validateJwtToken(validToken);

        Assertions.assertTrue(isValid);
    }

    @Test
    public void testExpiredToken() {
        JwtController jwtController = new JwtController();
        String expiredToken = generateExpiredToken();

        boolean isValid = jwtController.validateJwtToken(expiredToken);

        Assertions.assertFalse(isValid);
    }

    @Test
    public void testInvalidToken() {
        JwtController jwtController = new JwtController();
        String invalidToken = "invalidToken";

        boolean isValid = jwtController.validateJwtToken(invalidToken);

        Assertions.assertFalse(isValid);
    }

    private String generateValidToken() {
        // Generate a valid token for testing
        // ...

        return "validToken";
    }

    private String generateExpiredToken() {
        // Generate an expired token for testing
        // ...

        return "expiredToken";
    }
}